Refactor editor
